requirements.txt
python-dotenv~=1.0.0
pandas~=2.0.3
requests~=2.31.0
beautifulsoup4~=4.12.2
selenium~=4.10.0